requirements.txt
google-generativeai
plotly
pandas
streamlit